# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Basic Banking System - OpenAPI 3.0",
    "description": "Ini adalah dokumentasi menggunakan swagger untuk challenge 5 Basic Banking System Binar Academy x MSIB",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "Operations users"
    },
    {
      "name": "accounts",
      "description": "Operations accounts"
    },
    {
      "name": "transactions",
      "description": "Operations transactions"
    },
    {
      "name": "authentication",
      "description": "Operations authentication"
    }
  ],
  "paths": {
    "/api/v1/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Users"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create user",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by id",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts": {
      "get": {
        "tags": ["accounts"],
        "summary": "Get all accounts",
        "operationId": "getAccounts",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/bank_accounts"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["accounts"],
        "summary": "Create account",
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/bank_accounts"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/bank_accounts"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}": {
      "get": {
        "tags": ["accounts"],
        "summary": "Get account by id",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "ID of account to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/bank_accounts"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions": {
      "get": {
        "tags": ["transactions"],
        "summary": "Get all transactions",
        "operationId": "getTransactions",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/transactions"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["transactions"],
        "summary": "Create transaction",
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/transactions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/transactions"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "tags": ["transactions"],
        "summary": "Get transaction by id",
        "operationId": "getTransactionById",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "description": "ID of transaction to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/transactions"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["authentication"],
        "summary": "register user",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["authentication"],
        "summary": "login user",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                },
                "example": {
                  "email": "example@mail.com",
                  "password": "password"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login Successful",
            "content": {
              "application/json": {
                "example": {
                  "email": "example@mail.com",
                  "password": "password",
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwibmFtZSI6IkFkYW0iLCJlbWFpbCI6ImFkYW1AbWFpbC5jb20iLCJpYXQiOjE2NzYwNjYwNzJ9.1"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/authenticate": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["authentication"],
        "summary": "Check authenticate",
        "operationId": "authenticate",
        "responses": {
          "200": {
            "description": "Authenticated successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Authenticated successfully"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "format": "string",
            "example": "example"
          },
          "email": {
            "uniqueItems": true,
            "type": "string",
            "format": "string",
            "example": "example@mail.com"
          },
          "password": {
            "type": "string",
            "format": "string",
            "example": "password"
          }
        },
        "xml": {
          "name": "Users"
        }
      },
      "Profiles": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "user_id": {
            "uniqueItems": true,
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "identity_type": {
            "type": "string",
            "format": "string",
            "example": "KTP"
          },
          "identity_number": {
            "uniqueItems": true,
            "type": "string",
            "format": "string",
            "example": "1234567890"
          },
          "address": {
            "type": "string",
            "format": "string",
            "example": "jl. kenangan no. 123"
          }
        },
        "xml": {
          "name": "Profiles"
        }
      },
      "bank_accounts": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "user_id": {
            "uniqueItems": true,
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "bank_name": {
            "type": "string",
            "format": "string",
            "example": "BANK TOIB"
          },
          "bank_account_number": {
            "uniqueItems": true,
            "type": "string",
            "format": "string",
            "example": "9876543210"
          },
          "balance": {
            "type": "integer",
            "format": "int64",
            "example": 2000000
          }
        },
        "xml": {
          "name": "bank_accounts"
        }
      },
      "transactions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "source_account_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "destination_account_id": {
            "type": "integer",
            "format": "int64",
            "example": 2
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "example": 200000
          }
        },
        "xml": {
          "name": "transactions"
        }
      }
    }
  }
}